simplify filesystem file naming
We allowed pbf / mvt, which has overhead unless defined in a configuration
adapter - which results in more configuration.  We are NOW assuming pbf for
the moment as that is the most popular extension.

Note: for online endpoints, it is the TileJSON that defines the extension.
It is possible for us to introduce TileJSON for storage configuration but
that requires further thought.